Track the resulting filterlist in git
