+++ .gitignore
@@ -11,6 +11,5 @@ addon/*.js
 .vscode/
 .env
 .DS_Store
-rules/filterlist.txt
 rules/filterlists/easylist_*.txt
 lib/filterlist-engine.ts
