+++ package.json
@@ -20,8 +20,7 @@
     "clean": "rm -r dist",
     "lint": "eslint addon/ lib/ playwright/ tests/ tests-wtr/ rules/autoconsent/*.json",
     "bundle": "./build.sh",
-    "build-all": "npm run rebuild-filterlist && npm run build-rules && npm run bundle",
-    "watch": "npm run prepublish && chokidar \"lib\" \"addon\" \"rules/autoconsent\" \"rules/filterlists/*\" --ignore 'lib/filterlist-engine.ts' -c \"npm run build-all\"",
+    "watch": "npm run prepublish && chokidar \"lib\" \"addon\" \"rules/autoconsent\" \"rules/filterlist.txt\" --ignore 'lib/filterlist-engine.ts' -c \"npm run prepublish\"",
     "create-rule": "node rules/create-rule.mjs",
     "test": "playwright test",
     "test:webkit": "playwright test --project webkit",
@@ -29,13 +28,14 @@
     "test:chrome": "playwright test --project chrome",
     "test:lib": "web-test-runner",
     "test:sample": "playwright test tests/_sample-test.spec.ts --project webkit",
-    "fetch-easylist": "rm -f rules/filterlists/easylist_cookie_* && wget --directory-prefix=rules/filterlists/ https://raw.githubusercontent.com/easylist/easylist/refs/heads/master/easylist_cookie/easylist_cookie_general_hide.txt https://raw.githubusercontent.com/easylist/easylist/refs/heads/master/easylist_cookie/easylist_cookie_specific_hide.txt https://raw.githubusercontent.com/easylist/easylist/refs/heads/master/easylist_cookie/easylist_cookie_specific_uBO.txt https://raw.githubusercontent.com/easylist/easylist/refs/heads/master/easylist_cookie/easylist_cookie_international_specific_hide.txt https://raw.githubusercontent.com/easylist/easylist/refs/heads/master/easylist_cookie/easylist_cookie_allowlist_general_hide.txt https://raw.githubusercontent.com/easylist/easylist/refs/heads/master/easylist_cookie/easylist_cookie_allowlist.txt",
-    "rebuild-filterlist": "node rules/rebuild-filterlist.mjs",
-    "update-filterlist": "npm run fetch-easylist && npm run rebuild-filterlist",
+    "fetch-easylist": "scripts/fetch-easylist.sh",
+    "rebuild-filterlist": "node scripts/rebuild-filterlist.mjs",
+    "update-easylist": "npm run fetch-easylist && npm run rebuild-filterlist",
+    "compile-filterlist": "node scripts/compile-filterlist.mjs",
     "build-rules": "node rules/build.mjs",
     "version": "node update_version.js && git add addon/manifest.*.json",
     "vendor-copy": "mkdir -p addon/vendor && cp node_modules/mocha/mocha.* addon/vendor/ && cp node_modules/chai/chai.js addon/vendor/",
-    "prepublish": "npm run fetch-easylist && npm run build-all",
+    "prepublish": "npm run compile-filterlist && npm run build-rules && npm run bundle",
     "launch:firefox": "web-ext run -s dist/addon-firefox",
     "release": "auto shipit",
     "get-text-for-xpath": "ts-node scripts/get-text-for-xpath.ts"
